refactor: move functions getRNG, cyrb128, and sfc32 into utils.js
